requirements.txt
python-docx>=0.8.11
openpyxl>=3.0.9
Pillow>=9.0.0
